Disable Naming/RescuedExceptionsVariableName (#42)
In my opinion, the variable name for an exception is not important in
many cases. Some love to name it as `e`, some like `error`, some insist
on naming it as `exception`... I believe it absolutely a matter of
preference and does not undermine productivity in software development.
Besides, even if someone writes code with so an inappropriate variable,
this would be pointed out by other code reviewers.

So, I want to disable `Naming/RescuedExceptionsVariableName`.