Disable Style/WordArray (#92)
I don't think `Style/WordArray` is useful in our application.
In some cases, `%w` might be useful, but it sometimes causes
problems.

For example, `%w[true false]` is the same as `["true", "false"]`,
not `[true, false]`. In this case, I think `["true", "false"]` might be better.

So, I don't want to enforce this rule in a whole project.